+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ next version
 ------------
 
 * Add modeling.get_individuals
+* Add modeling.get_baselines
 
 0.40.0 (2021-12-16)
 -------------------